Backend/requirements.txt
Flask
gunicorn
requests
Pillow
flask-cors
Flask-CORS==3.1.1
python-dotenv==1.0.0
